Add FinancialDataService.Streams
This commit introduces a new service called FinancialDataService.Streams. The service is used to handle price updates and provide them to clients in real-time using SignalR. The service is receiving data from a TCP server for simple demonstration of the concept.  A real production ready system would ideally use a distributed backplane/broker such as redis or kafka.