Update package dependencies, change connection string, and update readme
This commit includes an update of package dependencies in multiple projects, the connection string has been changed in appsettings.json files, and the readme is updated with new setup instructions. Additional changes include executing a FetchInstrumentsJob initially to populate the database and fixes to the Dockerfile work directory path for FinancialDataService.Streams.